Iframe dialogs in IE8 can't handle history
The presence of iframes seem to confuse IE8's dealing with history. Since IE8 was using history.js's document.hash fallback instead of pushstate, it was sensitive to having the history disrupted. In this bug, history.back() would cause IE8 to navigate back 2 frames instead of 1 when dialogs were closed.