Added ModalDialog Plugin Support
Added support for modal dialog plugins to create custom dialog types.

$.modalDialog.registerPlugin accepts a callback that expects modalDialog
and initializers parameters. The ModalDialog parameter allows for
extending the prototype to create custom dialog types and the
intializers array can be prepended with a new function to create a
dialog of the custom type.

The former if/else construction logic is moved into the
_typeInitializers function array. If a function returns null, the next
one is called to try and create a dialog of that type.

Added unit tests for the plugin.