Add non-ajax Partial Load
Adding a public function similar to partialLoad, which takes a string of
html rather than a url. It skips the ajax request entirely. The other
primary difference is that it does not load the javascript itself, it
passes the script tags back through an array. The caller can decide
whether to load them or not (there are specific files we need to avoid
loading).